Decouple smsOptInForm component from success screen
